Clean Code review and New Toolbar component
