+++ bogota-abril-2019/final_project/marvin-cely/tenis-match/src/App.js
@@ -99,7 +99,6 @@ class App extends React.Component {
           <Router>
             {navBar()}
             <section>
-            {this.props.store.tournaments.length}
               {routesList}
             </section>
           </Router>
