+++ bogota-abril-2019/final_project/marvin-cely/tenis-match/src/components/TournamentPageComponent/TournamentLogo.js
@@ -1,23 +1,27 @@
 import React from 'react';
 import '../../css/TournamentLogo.css';
 import playingTennis from '../../images/playingTennis.png';
+import { observer } from 'mobx-react';
 
-class TournamentLogo extends React.Component {
-  constructor(props) {
-    super(props);
-  }
 
-  render() { 
-    return(
-      <React.Fragment>
-        <figure aria-label='Logo del torneo'>
-          <h2>{this.props.maxScore}</h2>
-          <img src={playingTennis} alt="Torneo Logo" className="img-fluid"></img>
-          <h1>{this.props.tournamentName}</h1>
-        </figure>
-      </React.Fragment>
-    );
+const TournamentLogo = observer(
+  class TournamentLogo extends React.Component {
+    constructor(props) {
+      super(props);
+    }
+
+    render() { 
+      return(
+        <React.Fragment>
+          <figure aria-label='Logo del torneo'>
+            <h2>{this.props.maxScore}</h2>
+            <img src={playingTennis} alt="Torneo Logo" className="img-fluid"></img>
+            <h1>{this.props.tournamentName}</h1>
+          </figure>
+        </React.Fragment>
+      );
+    }
   }
-}
+);
 
 export default TournamentLogo;
