+++ bogota-abril-2019/final_project/marvin-cely/tenis-match/src/components/TournamentPageComponent/TournamentsPage.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import '../../css/TournamentsPage.css';
 import { observer } from 'mobx-react';
+import TournamentPageTools from './TournamentPageTools';
 import TournamentCard from './TournamentCard';
 import { LoadingComponent, ErrorServiceComponent, BackgroundImage } from '../SmallPieceComponent';
 import { dataTournamentPage } from '../../data-component/data-tournament-page';
@@ -12,17 +13,42 @@ const TournamentsPage = observer(
   class TournamentsPage extends React.Component {
     constructor(props) {
       super(props);
+      
       this.state = {
         bannerBackgroundDescription: dataTournamentPage.bannerBackgroundDescription,
       }
+
+      this.handleInput = this.handleInput.bind(this);
+    }
+
+    handleInput = (event) => {
+      const { name, value } = event.target;
+      // if (name === 'editPlayer' || name === 'addPlayer') {
+      //   this.setState({
+      //     idPlayerSelected: value,
+      //     counterCRUDAction: this.state.counterCRUDAction + 1,
+      //   });
+      // } 
+      // else {
+      //   if(name === 'toggleOrder') {     
+      //     this.props.store.playersTable.isAscending = !this.props.store.playersTable.isAscending;
+      //   }
+      //   else if (name === 'orderOption') {
+      //     this.props.store.playersTable.orderType = value;
+      //   }
+      //   else if (name === 'search') {
+      //     this.props.store.playersTable.searchValue = value;
+      //   }
+      //   this.setState({ counterTableAction: this.state.counterTableAction + 1 });
+      // }
     }
 
     render() { 
       const tournamentContent = () => {
         return(
-          <section>
-            {this.props.store.tournaments.length}
-          </section>        
+          <main className="container"> 
+            <TournamentPageTools handleInput={this.handleInput} store={this.props.store}/>
+          </main>        
         );
       }
