+++ bogota-abril-2019/final_project/marvin-cely/tenis-match/src/css/MatchesPage.css
@@ -1,4 +1,4 @@
-figure[data-background-name^="Puntua"] {
+figure[data-background-name="Partidos"] {
     background-image: url("../images/puntuacionImagen.jpg");
     min-height: 40vh; 
     background-attachment: fixed;
@@ -11,7 +11,7 @@ figure[data-background-name^="Puntua"] {
     margin: 0;
 }
 
-figure[data-background-name^="Puntua"] > figcaption {
+figure[data-background-name="Partidos"] > figcaption {
     padding: 2vh;
     font-size: var(--size-large);
     font-family: var(--font-group-3);
