+++ bogota-abril-2019/final_project/marvin-cely/tenis-match/src/css/PlayerPageTools.css
@@ -1,0 +1,91 @@
+.btn-group, header section > button {
+    margin: 0.8vh;
+}
+
+main.container > section:first-of-type {
+    z-index: 3;
+}
+
+@media only screen and (min-width: 992px) {
+    section > header.row > section:first-of-type  {
+        margin-bottom: 1.2rem;
+    }
+    main.container > section:first-of-type {
+        padding-top: 4rem;
+        margin-top: -4rem;
+    }
+}
+
+@media only screen and (max-width: 991px) {
+    section > header.row > section:first-of-type section:first-of-type  {
+        margin-bottom: 0.3rem;
+    }
+    main.container > section:first-of-type {
+        padding-top: 3.6rem;
+        margin-top: -3.6rem;
+    }
+}
+
+form > fieldset {
+    margin-bottom: 0;
+}
+
+section > header.row {
+    background: white;
+    padding: 1rem;
+    padding-bottom: 0.8rem;
+    border-bottom: 1px solid var(--font-color-gray-light);
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    flex-flow: row wrap;
+}
+
+@media only screen and (max-width: 767px) {
+    section > header.row {
+        box-shadow: 0 1.5rem 1rem -1rem rgba(0, 0, 0, .09); 
+    }
+}
+
+section > header.row > section:last-of-type, section > header.row > section:first-of-type > div section {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    flex-direction: column;
+}
+
+section > header.row > section:first-of-type > div > section:last-of-type div[data-toggle="buttons"]{
+    margin: 0;
+}
+
+header > section:last-of-type + form {
+    margin: 1vh;
+}
+
+header > section:last-of-type h5 {
+    text-align: center;
+    margin-bottom: 0;
+}
+
+section.row {
+    margin: auto;
+    text-align: center;
+    padding: 1rem;;
+}
+
+section > header.row > section:first-of-type > div > section:first-of-type [name='search'] {
+    width: 19.5rem;
+    padding-left: 2.375rem;
+}
+
+section > header.row > section:first-of-type > div > section:first-of-type span {
+    position: absolute;
+    z-index: 2;
+    display: block;
+    width: 2.375rem;
+    height: 2.375rem;
+    line-height: 2.375rem;
+    text-align: center;
+    pointer-events: none;
+    color: var(--font-color-gray);
+}
