--- bogota-abril-2019/final_project/marvin-cely/tenis-match/src/data-component/data-scores-page.json
@@ -1,5 +1,0 @@
-{
-    "dataScoresPage": {
-        "bannerBackgroundDescription":""
-    }
-}
